pothole_detection_app/requirements.txt
flask==2.3.3
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0# Requirements